pyproject.toml
[project]
name = "livekit-voice-agent"
version = "0.1.0"
description = "LiveKit-powered voice AI agent with MCP server integration"
requires-python = ">=3.9"
dependencies = [
    "livekit>=0.17.0",
    "livekit-agents>=0.9.0",
    "livekit-plugins-openai>=0.7.0",
    "livekit-plugins-deepgram>=0.6.0",
    "livekit-plugins-silero>=0.6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = []

[tool.uv]
package = false

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"
